refactor(pipeline): split meta_analysis.py aggregations into helper module (#662)
apps/pipeline/app/services/meta_analysis.py was 565 LOC. The per-feed,
per-episode, per-speaker, timeline, and coverage builders move into a
new meta_analysis_aggregations module along with the tiktoken setup and
the _count_tokens / _count_turns helpers they consume. The orchestrator
module keeps the public entry points (is_stale / set_stale / clear_stale
/ compute_snapshot / upsert_snapshot / recompute_and_store) plus the
stale-flag token helpers, and re-exports every moved symbol via __all__
so existing call sites (app.tasks.infer, app.tasks.archive, app.worker,
app.api.meta_analysis) and the existing test suite (which references
e.g. svc._count_turns through the orchestrator module) keep working
unchanged. No behavior changes.
